added interface for each service
