+++ DairyStore/src/main/java/com/dairystore/Controllers/CartController.java
@@ -3,7 +3,6 @@ package com.dairystore.Controllers;
 import com.dairystore.Models.dtos.ShoppingCartDto;
 import com.dairystore.Services.CartItemService;
 import com.dairystore.Services.CartService;
-import com.dairystore.Services.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +13,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
 
@@ -24,7 +22,6 @@ import java.util.List;
 public class CartController {
     private final CartService cartService;
     private final CartItemService cartItemService;
-    private final UserService userService;
 
     @PostMapping("/add")
     public ResponseEntity<String> addToCart(@RequestParam Long productId, @RequestParam int quantity) {
@@ -41,12 +38,6 @@ public class CartController {
         return ResponseEntity.ok().body(cartService.viewShoppingCart());
     }
 
-    /*@PostMapping("/deleteById")
-    public String deleteProductById(@RequestParam("productId") Long productId, RedirectAttributes redirectAttributes) {
-        cartItemService.deleteCartItemsByProductId(productId);
-        redirectAttributes.addFlashAttribute("success", "Product is successfully deleted");
-        return "redirect:/test/buyer";
-    }*/
     @DeleteMapping("/{productId}")
     public ResponseEntity<?> deleteProductById(@PathVariable Long productId) {
         try {
