+++ DairyStore/src/main/java/com/dairystore/Services/CartServiceImpl.java
@@ -1,0 +1,90 @@
+package com.dairystore.Services;
+
+import com.dairystore.Models.Cart;
+import com.dairystore.Models.CartItem;
+import com.dairystore.Models.Product;
+import com.dairystore.Models.User;
+import com.dairystore.Models.dtos.ShoppingCartDto;
+import com.dairystore.Repository.CartRepository;
+import lombok.RequiredArgsConstructor;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+public class CartServiceImpl implements CartService {
+    private final CartRepository cartRepository;
+    private final UserServiceImpl userService;
+    private final ProductServiceImpl productService;
+    private final CartItemServiceImpl cartItemService;
+
+    @Override
+    public void addToCart(Long productId, int quantity) throws Exception {
+        User user = userService.getUserByUsername();
+        Product product = productService.getProductById(productId);
+        validateProductQuantity(quantity, product);
+        Cart cart = findOrCreateCart(user);
+        cartItemService.saveCartItems(cart, product, quantity);
+    }
+
+    private void validateProductQuantity(int quantity, Product product) throws Exception {
+        if (quantity > product.getQuantity()) {
+            throw new Exception("Няма достатъчно количество от продукта в наличност!");
+        }
+    }
+
+    @NotNull
+    private Cart findOrCreateCart(User user) {
+        Cart cart = cartRepository.findByUser(user);
+        if (cart == null) {
+            cart = new Cart();
+            cart.setUser(user);
+            cartRepository.save(cart);
+        }
+        return cart;
+    }
+
+    @Override
+    public List<ShoppingCartDto> viewShoppingCart() {
+        User user = userService.getUserByUsername();
+        List<CartItem> cartItemList = cartItemService.getCartItemsByCart(user.getCart());
+
+        return cartItemList.stream()
+                .map(cartItem -> mapToShoppingCartDto(user, cartItem))
+                .collect(Collectors.toList());
+    }
+
+    private ShoppingCartDto mapToShoppingCartDto(User user, CartItem cartItem) {
+        Product product = cartItem.getProduct();
+        int quantity = cartItem.getQuantity();
+        double discount = user.getCompanyName().isEmpty() ? 0 : product.getDiscount();
+        double totalPricePerProduct = calculateTotalPricePerProduct(user, product, quantity);
+
+        return ShoppingCartDto.builder()
+                .id(product.getId())
+                .name(product.getName())
+                .type(product.getType())
+                .weight(product.getWeight())
+                .discount(discount)
+                .price(product.getPrice())
+                .quantity(quantity)
+                .totalPricePerProduct(totalPricePerProduct)
+                .build();
+    }
+
+    private double calculateTotalPricePerProduct(User user, Product product, int quantity) {
+        double basePrice = product.getPrice() * quantity;
+        double discountAmount = user.getCompanyName().isEmpty() ? 0 : basePrice * (product.getDiscount() / 100);
+        double total = basePrice - discountAmount;
+        return Math.round(total * 100.0) / 100.0;
+    }
+
+    @Override
+    public Cart getCartByUser(User user) {
+        return cartRepository.findByUser(user);
+    }
+
+}
