+++ DairyStore/src/main/java/com/dairystore/Services/DeliveryCompanyService.java
@@ -2,49 +2,18 @@ package com.dairystore.Services;
 
 import com.dairystore.Models.DeliveryCompany;
 import com.dairystore.Models.dtos.CreateDeliveryCompanyDto;
-import com.dairystore.Repository.DeliveryCompanyRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import org.springframework.validation.BindingResult;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class DeliveryCompanyService {
-    private final DeliveryCompanyRepository deliveryCompanyRepository;
+public interface DeliveryCompanyService {
+    List<DeliveryCompany> getDeliveryCompanies();
 
-    public List<DeliveryCompany> getDeliveryCompanies() {
-        return deliveryCompanyRepository.findAll();
-    }
+    DeliveryCompany getDeliveryCompanyById(Long companyId);
 
-    public DeliveryCompany getDeliveryCompanyById(Long companyId) {
-        return deliveryCompanyRepository.findDeliveryCompanyById(companyId);
-    }
+    void deleteDeliveryCompanyById(Long deliveryCompanyId) throws Exception;
 
-    public void deleteDeliveryCompanyById(Long deliveryCompanyId) throws Exception {
-        isExist(deliveryCompanyId);
-        deliveryCompanyRepository.deleteById(deliveryCompanyId);
-    }
+    void createDeliveryCompany(CreateDeliveryCompanyDto createDeliveryCompanyDto);
 
-    private void isExist(Long deliveryCompanyId) throws Exception {
-        if (!deliveryCompanyRepository.existsById(deliveryCompanyId)) {
-            throw new Exception("Фирмата не съществува в базата данни");
-        }
-    }
-
-    public void createDeliveryCompany(CreateDeliveryCompanyDto createDeliveryCompanyDto) {
-        DeliveryCompany deliveryCompany = DeliveryCompany.builder()
-                .name(createDeliveryCompanyDto.getName())
-                .deliveryFee(createDeliveryCompanyDto.getDeliveryFee())
-                .build();
-
-        deliveryCompanyRepository.save(deliveryCompany);
-    }
-
-    public void isExistCompanyByName(String name, BindingResult bindingResult) {
-        if (deliveryCompanyRepository.existsByName(name)) {
-            bindingResult.rejectValue("name", "error.name", "Името на фирмата вече съществува");
-        }
-    }
+    void isExistCompanyByName(String name, BindingResult bindingResult);
 }
