+++ DairyStore/src/main/java/com/dairystore/Services/analyses/buyer/BuyerAnalysisService.java
@@ -1,23 +1,7 @@
 package com.dairystore.Services.analyses.buyer;
 
-import com.dairystore.Models.User;
 import com.dairystore.Models.dtos.BuyerAnalyticsDto;
-import com.dairystore.Services.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class BuyerAnalysisService {
-    private final UserService userService;
-    private final List<BuyerAnalysisStrategy> strategyList;
-
-    public BuyerAnalyticsDto getBuyerAnalysis() {
-        User user = userService.getUserByUsername();
-        BuyerAnalyticsDto.BuyerAnalyticsDtoBuilder builder = BuyerAnalyticsDto.builder();
-        strategyList.forEach(strategyList -> strategyList.analyze(user, builder));
-        return builder.build();
-    }
+public interface BuyerAnalysisService {
+    BuyerAnalyticsDto getBuyerAnalysis();
 }
