+++ DairyStore/src/main/java/com/dairystore/Services/analyses/buyer/BuyerAnalysisServiceImpl.java
@@ -1,0 +1,24 @@
+package com.dairystore.Services.analyses.buyer;
+
+import com.dairystore.Models.User;
+import com.dairystore.Models.dtos.BuyerAnalyticsDto;
+import com.dairystore.Services.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class BuyerAnalysisServiceImpl implements BuyerAnalysisService {
+    private final UserService userService;
+    private final List<BuyerAnalysisStrategy> strategyList;
+
+    @Override
+    public BuyerAnalyticsDto getBuyerAnalysis() {
+        User user = userService.getUserByUsername();
+        BuyerAnalyticsDto.BuyerAnalyticsDtoBuilder builder = BuyerAnalyticsDto.builder();
+        strategyList.forEach(strategyList -> strategyList.analyze(user, builder));
+        return builder.build();
+    }
+}
